Restructure project to cmd/ and pkg/
