+++ Dockerfile
@@ -5,7 +5,7 @@ RUN go mod download
 COPY . .
 RUN --mount=type=cache,target=/go/pkg/mod \
     --mount=type=cache,target=/root/.cache/go-build \
-    CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o kubespiffed .
+    CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ./kubespiffed cmd/kubespiffe/main.go
 FROM scratch
 COPY --from=builder /build/kubespiffed /app/
 ENTRYPOINT ["/app/kubespiffed"]
