+++ Justfile
@@ -1,6 +1,5 @@
 build:
-	GOOS=linux GOARCH=amd64 go build -o kubespiffed
-
+	GOOS=linux GOARCH=amd64 go build -o ./kubespiffed cmd/kubespiffe/main.go
 gen verb='':
 	#!/usr/bin/env bash
 	if [ "{{verb}}" = "verify" ]; then
