+++ cmd/kubespiffe/main.go
@@ -6,6 +6,8 @@ import (
 	"log/slog"
 	"net/http"
 	"os"
+
+	"github.com/jsnctl/kubespiffe/pkg/k8s"
 )
 
 const (
@@ -14,34 +16,34 @@ const (
 
 func main() {
 	ctx := context.Background()
-	cs, err := getKubernetesClientset()
+	cs, err := k8s.GetKubernetesClientset()
 	if err != nil {
 		log.Fatalf("problem with k8s clientset: %v", err)
 	}
-	kscs, err := getKubespiffeClientset()
+	kscs, err := k8s.GetKubespiffeClientset()
 	if err != nil {
 		log.Fatalf("problem with kubespiffe clientset: %v", err)
 	}
 	http.HandleFunc("/v1/svid", func(w http.ResponseWriter, r *http.Request) {
-		token := extractBearerToken(r.Header.Get("Authorization"))
+		token := k8s.ExtractBearerToken(r.Header.Get("Authorization"))
 		if token == "" {
 			http.Error(w, "missing bearer token", http.StatusUnauthorized)
 			return
 		}
 
-		jwks, err := getKubernetesJWKS(ctx)
+		jwks, err := k8s.GetKubernetesJWKS(ctx)
 		if err != nil {
 			slog.Error("problem with JWKS")
 			return
 		}
 
-		claims, err := verifyPSAT(token, jwks)
+		claims, err := k8s.VerifyPSAT(token, jwks)
 		if err != nil {
 			slog.Error("problem with PSAT", "error", err)
 			return
 		}
 
-		if err := attestPod(ctx, cs, kscs, claims["kubernetes.io"].(map[string]any)); err != nil {
+		if err := k8s.AttestPod(ctx, cs, kscs, claims["kubernetes.io"].(map[string]any)); err != nil {
 			slog.Info("❌ Pod rejected", "error", err)
 		}
 		w.Header().Set("Content-Type", "application/json")
