+++ pkg/k8s/helpers_test.go
@@ -1,4 +1,4 @@
-package main
+package k8s
 
 import (
 	"crypto/rand"
@@ -45,7 +45,7 @@ func Test_extractBearer(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := extractBearerToken(tt.header)
+			got := ExtractBearerToken(tt.header)
 			assert.Equal(t, got, tt.want)
 		})
 	}
